Encoded Alouette: rhyme analysis not done
there was some trouble with this and I'm not sure whether I've represented the verses correctly, or if it would be better for them to have independent spines from the beginning. Also, the segno repeats/structure is difficult to represent -should I represent the enumeration at the top-level structure as well? Is there a way to create a substructure to represent it properly?